changes for proto

+++ src/main/java/com/sjsu/rollbits/datasync/server/resources/GroupResource.java
@@ -55,7 +55,7 @@ public class GroupResource implements RouteResource {
 				// save to database
 
 				for (RNode node : nodes) {
-					MessageClient mc = new MessageClient(node.getIpAddress(), node.getPort());
+					MessageClient mc = new MessageClient(node.getIpAddress(),(int) node.getPort());
 					if (node.getType().equals(RNode.Type.REPLICA)) {
 						mc.addGroup(group.getGname(), (int) group.getGid(), true, true);
 					} else {
